Update and rename otaUpdate.h to WiFiClientSecure.h
Refactored the ota updates into a separate class. Leaving the wifi client call (get request) And renamed this file into callServer_WiFiClientSecure.h
